+++ src/Humans.Web/Views/Admin/Index.cshtml
@@ -17,14 +17,14 @@
                 <a asp-action="Configuration" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-gear me-2"></i>Configuration Status
                 </a>
-                <a asp-action="EmailPreview" class="list-group-item list-group-item-action">
+                <a asp-controller="Email" asp-action="EmailPreview" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-envelope me-2"></i>Email Preview
                 </a>
-                <a asp-action="EmailOutbox" class="list-group-item list-group-item-action">
+                <a asp-controller="Email" asp-action="EmailOutbox" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-inbox me-2"></i>Email Outbox
                 </a>
-                <a asp-action="SyncSettings" class="list-group-item list-group-item-action">
-                    <i class="fa-solid fa-sliders me-2"></i>Sync Settings
+                <a asp-controller="Google" asp-action="Index" class="list-group-item list-group-item-action">
+                    <i class="fa-brands fa-google me-2"></i>Google Integration
                 </a>
                 <a href="/hangfire" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-clock-rotate-left me-2"></i>@Localizer["Admin_BackgroundJobs"]
@@ -41,7 +41,7 @@
                 <a asp-action="Logs" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-triangle-exclamation me-2"></i>Application Logs
                 </a>
-                <a asp-controller="AdminEmail" asp-action="Index" class="list-group-item list-group-item-action">
+                <a asp-controller="Google" asp-action="Accounts" class="list-group-item list-group-item-action">
                     <i class="fa-solid fa-at me-2"></i>@@nobodies.team Accounts
                 </a>
                 <a asp-controller="AdminMerge" asp-action="Index" class="list-group-item list-group-item-action">
@@ -50,56 +50,6 @@
             </div>
         </div>
 
-        <div class="card mb-4">
-            <div class="card-header">System Team Sync</div>
-            <div class="card-body">
-                <form asp-action="SyncSystemTeams" method="post">
-                    @Html.AntiForgeryToken()
-                    <button type="submit" class="btn btn-outline-primary">
-                        <i class="fa-solid fa-arrows-rotate me-1"></i>Sync System Teams
-                    </button>
-                    <small class="text-muted d-block mt-1">Recalculates Volunteers, Leads, and Board team membership.</small>
-                </form>
-                <a asp-action="SyncResults" class="btn btn-outline-secondary btn-sm mt-3">
-                    <i class="fa-solid fa-list-check me-1"></i>View Sync Results
-                </a>
-            </div>
-        </div>
-
-        <div class="card mb-4">
-            <div class="card-header">Google Group Settings</div>
-            <div class="card-body">
-                <form asp-action="CheckGroupSettings" method="post">
-                    @Html.AntiForgeryToken()
-                    <button type="submit" class="btn btn-outline-primary">
-                        <i class="fa-solid fa-magnifying-glass me-1"></i>Check Group Settings
-                    </button>
-                    <small class="text-muted d-block mt-1">Checks all Google Groups for settings drift. Does not modify anything.</small>
-                </form>
-                <a asp-action="GroupSettingsResults" class="btn btn-outline-secondary btn-sm mt-3">
-                    <i class="fa-solid fa-list-check me-1"></i>View Last Group Settings Results
-                </a>
-                <a asp-action="AllGroups" class="btn btn-outline-primary btn-sm mt-2">
-                    <i class="fa-solid fa-globe me-1"></i>All Domain Groups
-                </a>
-            </div>
-        </div>
-
-        <div class="card mb-4">
-            <div class="card-header">Email Backfill</div>
-            <div class="card-body">
-                <form asp-action="CheckEmailMismatches" method="post">
-                    @Html.AntiForgeryToken()
-                    <button type="submit" class="btn btn-outline-primary">
-                        <i class="fa-solid fa-envelope-circle-check me-1"></i>Check Email Mismatches
-                    </button>
-                    <small class="text-muted d-block mt-1">Compares stored emails against Google Admin SDK. Detects case differences in @@nobodies.team addresses.</small>
-                </form>
-                <a asp-action="EmailBackfillReview" class="btn btn-outline-secondary btn-sm mt-3">
-                    <i class="fa-solid fa-list-check me-1"></i>View Last Results
-                </a>
-            </div>
-        </div>
     </div>
 
     <div class="col-md-6">
