+++ src/Humans.Web/Views/Google/EmailBackfillReview.cshtml
@@ -1,0 +1,93 @@
+@model Humans.Application.DTOs.EmailBackfillResult
+@{
+    ViewData["Title"] = "Email Backfill Review";
+}
+
+<div class="container-fluid px-4">
+    <div class="d-flex justify-content-between align-items-center mt-4 mb-3">
+        <h1>Email Backfill Review</h1>
+        <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
+            <i class="fa-solid fa-arrow-left me-1"></i>Back to Google
+        </a>
+    </div>
+
+    @if (!string.IsNullOrEmpty(Model.ErrorMessage))
+    {
+        <div class="alert alert-danger">
+            <i class="fa-solid fa-triangle-exclamation me-1"></i>Error: @Model.ErrorMessage
+        </div>
+    }
+    else if (Model.Mismatches.Count == 0)
+    {
+        <div class="alert alert-success">
+            <i class="fa-solid fa-check-circle me-1"></i>All @Model.TotalUsersChecked user(s) have matching emails. No corrections needed.
+        </div>
+    }
+    else
+    {
+        <div class="alert alert-warning">
+            <i class="fa-solid fa-triangle-exclamation me-1"></i>
+            Found <strong>@Model.Mismatches.Count</strong> email mismatch(es) out of <strong>@Model.TotalUsersChecked</strong> users checked.
+            Select the corrections to apply and click <strong>Apply Selected Corrections</strong>.
+        </div>
+
+        <form asp-action="ApplyEmailBackfill" method="post">
+            @Html.AntiForgeryToken()
+
+            <div class="card mb-4">
+                <div class="card-header d-flex justify-content-between align-items-center">
+                    <span><i class="fa-solid fa-envelope me-1"></i>Email Mismatches</span>
+                    <div class="form-check form-check-inline mb-0">
+                        <input class="form-check-input" type="checkbox" id="selectAll">
+                        <label class="form-check-label" for="selectAll">Select all</label>
+                    </div>
+                </div>
+                <div class="card-body p-0">
+                    <table class="table table-sm table-hover mb-0">
+                        <thead>
+                            <tr>
+                                <th class="ps-3" style="width:40px;"></th>
+                                <th>Human</th>
+                                <th>Stored Email</th>
+                                <th>Google Email</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            @foreach (var mismatch in Model.Mismatches)
+                            {
+                                <tr>
+                                    <td class="ps-3">
+                                        <input class="form-check-input row-checkbox"
+                                               type="checkbox"
+                                               name="selectedUserIds"
+                                               value="@mismatch.UserId" />
+                                        <input type="hidden"
+                                               name="corrections[@mismatch.UserId]"
+                                               value="@mismatch.GoogleEmail" />
+                                    </td>
+                                    <td>@mismatch.DisplayName</td>
+                                    <td><code>@mismatch.StoredEmail</code></td>
+                                    <td>
+                                        <code class="text-success">@mismatch.GoogleEmail</code>
+                                        <i class="fa-solid fa-arrow-right ms-1 text-muted"></i>
+                                    </td>
+                                </tr>
+                            }
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+
+            <button type="submit" class="btn btn-primary"
+                    data-confirm="Apply selected email corrections? This will update the stored email, username, and normalized fields for each selected user.">
+                <i class="fa-solid fa-check me-1"></i>Apply Selected Corrections
+            </button>
+        </form>
+    }
+</div>
+
+<script>
+    document.getElementById('selectAll')?.addEventListener('change', function () {
+        document.querySelectorAll('.row-checkbox').forEach(cb => cb.checked = this.checked);
+    });
+</script>
