+++ src/Humans.Web/Views/Google/SyncResults.cshtml
@@ -1,0 +1,76 @@
+@model Humans.Application.DTOs.SyncReport
+@{
+    ViewData["Title"] = "System Team Sync Results";
+}
+
+<div class="container-fluid px-4">
+    <div class="d-flex justify-content-between align-items-center mt-4 mb-3">
+        <h1>Sync Results</h1>
+        <a asp-action="Index" class="btn btn-outline-secondary btn-sm">
+            <i class="fa-solid fa-arrow-left me-1"></i>Back to Google
+        </a>
+    </div>
+
+    @if (!Model.HasChanges)
+    {
+        <div class="alert alert-success">
+            <i class="fa-solid fa-check-circle me-1"></i>All @Model.Steps.Count steps completed — no changes needed.
+        </div>
+    }
+    else
+    {
+        <div class="alert alert-info">
+            <i class="fa-solid fa-arrows-rotate me-1"></i>Sync completed with @Model.TotalChanges change(s) across @Model.Steps.Count steps.
+        </div>
+    }
+
+    @foreach (var step in Model.Steps)
+    {
+        <div class="card mb-3">
+            <div class="card-header d-flex justify-content-between align-items-center">
+                <strong>@step.StepName</strong>
+                @if (step.HasChanges)
+                {
+                    <span class="badge bg-warning text-dark">@step.TotalChanges change(s)</span>
+                }
+                else
+                {
+                    <span class="badge bg-success">No changes</span>
+                }
+            </div>
+            @if (step.HasChanges)
+            {
+                <div class="card-body p-0">
+                    <table class="table table-sm mb-0">
+                        <thead>
+                            <tr>
+                                <th style="width: 100px">Action</th>
+                                <th>Human</th>
+                                <th>Detail</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            @foreach (var change in step.Changes)
+                            {
+                                var badgeClass = change.Action switch
+                                {
+                                    "Added" => "bg-success",
+                                    "Removed" => "bg-danger",
+                                    "Fixed" => "bg-info",
+                                    _ => "bg-secondary"
+                                };
+                                <tr>
+                                    <td><span class="badge @badgeClass">@change.Action</span></td>
+                                    <td>
+                                        <human-link user-id="@change.User.UserId" display-name="@change.User.DisplayName" admin="true" show-popover="true" />
+                                    </td>
+                                    <td class="text-muted">@change.Detail</td>
+                                </tr>
+                            }
+                        </tbody>
+                    </table>
+                </div>
+            }
+        </div>
+    }
+</div>
